pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pii-cleaner"
version = "0.1.0"
description = "Production-grade PII cleaning service"
requires-python = ">=3.12"
license = { text = "Proprietary" }
readme = "README.md"
dependencies = [
    "fastapi>=0.115,<0.116",
    "uvicorn[standard]>=0.32,<0.33",
    "gunicorn>=23.0,<24.0",
    "pydantic>=2.9,<3.0",
    "pydantic-settings>=2.6,<3.0",
    "presidio-analyzer>=2.2.355,<3.0",
    "presidio-anonymizer>=2.2.355,<3.0",
    "spacy>=3.7,<4.0",
    "structlog>=24.4,<25.0",
    "prometheus-client>=0.21,<0.22",
    "argon2-cffi>=23.1,<24.0",
    "pyarrow>=18.0,<19.0",
    "pyyaml>=6.0,<7.0",
    "python-ulid>=3.0,<4.0",
    "opentelemetry-sdk>=1.28,<2.0",
    "opentelemetry-exporter-otlp>=1.28,<2.0",
    "opentelemetry-instrumentation-fastapi>=0.49b0,<1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3,<9.0",
    "pytest-asyncio>=0.24,<0.25",
    "pytest-cov>=6.0,<7.0",
    "httpx>=0.27,<0.28",
    "hypothesis>=6.115,<7.0",
    "schemathesis>=3.38,<4.0",
    "ruff>=0.7,<0.8",
    "black>=24.10,<25.0",
    "mypy>=1.13,<2.0",
    "pip-audit>=2.7,<3.0",
    "types-pyyaml",
]

[tool.hatch.build.targets.wheel]
packages = ["src/pii_cleaner"]

[tool.ruff]
line-length = 100
target-version = "py312"
src = ["src", "tests"]

[tool.ruff.lint]
select = [
    "E", "F", "W",    # pycodestyle, pyflakes
    "I",               # isort
    "B",               # bugbear
    "UP",              # pyupgrade
    "S",               # bandit
    "C4",              # comprehensions
    "SIM",             # simplify
    "RUF",             # ruff-specific
]
ignore = [
    "S101",  # assert allowed in tests
]

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S"]

[tool.black]
line-length = 100
target-version = ["py312"]

[tool.mypy]
python_version = "3.12"
strict = true
plugins = ["pydantic.mypy"]
packages = ["pii_cleaner"]
mypy_path = "src"

[[tool.mypy.overrides]]
module = ["presidio_analyzer.*", "presidio_anonymizer.*", "pyarrow.*"]
ignore_missing_imports = true

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
addopts = "-ra --strict-markers --strict-config"
filterwarnings = ["error"]

[tool.coverage.run]
branch = true
source = ["src/pii_cleaner"]

[tool.coverage.report]
fail_under = 80
show_missing = true
skip_covered = true
exclude_lines = [
    "pragma: no cover",
    "raise NotImplementedError",
    "if TYPE_CHECKING:",
]
